requirements.txt
annotated-types==0.6.0
anyio==4.3.0
article-parser==1.7.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.3
googlesearch-python==1.2.4
h11==0.14.0
html2text==2020.1.16
idna==3.7
lxml==5.1.1
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
